+++ src/models/tree/node.js
@@ -27,7 +27,6 @@ function fwc_tree_node(node_data) {
 		this.id = node_data.id;
 		this.text = node_data.name;
 		this.pid = node_data.id_parent;
-		this.allowdrag = 0;
 		this.node_type = node_data.node_type;
 		this.obj_type = node_data.obj_type;
 		this.id_obj = node_data.id_obj;